feat: persisted RPS round results

+++ gamemaster/db/datasets/rps_result_dataset.py
@@ -1,4 +1,4 @@
-from peewee import DateTimeField, ForeignKeyField, IntegerField
+from peewee import BigIntegerField, DateTimeField, ForeignKeyField
 
 from ..base import BaseModel
 from ..datasets import PlayerDataset
@@ -7,5 +7,5 @@ from ..datasets import PlayerDataset
 class RPSResultDataset(BaseModel):
     player_1 = ForeignKeyField(PlayerDataset, backref="rps_games_as_first")
     player_2 = ForeignKeyField(PlayerDataset, backref="rps_games_as_second")
-    rounds = IntegerField()
+    rounds = BigIntegerField()
     saved = DateTimeField()
