+++ gamemaster/db/migrations/20260526010543_add_rps_result.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING, Any
 
-from peewee import DateTimeField, ForeignKeyField, IntegerField
+from peewee import BigIntegerField, DateTimeField, ForeignKeyField
 
 from gamemaster.db.base import BaseModel
 from gamemaster.db.datasets import PlayerDataset
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
 class RPSResultDataset(BaseModel):
     player_1 = ForeignKeyField(PlayerDataset, backref="rps_games_as_first")
     player_2 = ForeignKeyField(PlayerDataset, backref="rps_games_as_second")
-    rounds = IntegerField()
+    rounds = BigIntegerField()
     saved = DateTimeField()
